pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nblite"
version = "0.1.0"
description = "Notebook-driven Python package development tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Lukas Kikuchi", email = "lukas@example.com" }
]
keywords = ["jupyter", "notebook", "nbdev", "literate-programming", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "notebookx-py>=0.1.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
docs = [
    "jupyter-book>=0.15.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocs-jupyter>=0.24.0",
]

[project.scripts]
nbl = "nblite.cli:main"

[project.urls]
Homepage = "https://github.com/lukastk/nblite"
Documentation = "https://github.com/lukastk/nblite"
Repository = "https://github.com/lukastk/nblite"

[tool.hatch.build.targets.wheel]
packages = ["nblite"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["nblite"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
